Fixed type hint of `timeout` parameter to `_wait_for_data()`

+++ adafruit_as7341.py
@@ -413,7 +413,7 @@ class AS7341:  # pylint:disable=too-many-instance-attributes, no-member
         self._configure_f5_f8()
         return self._channel_5_data
 
-    def _wait_for_data(self, timeout: Optional[int] = 1.0) -> None:
+    def _wait_for_data(self, timeout: Optional[float] = 1.0) -> None:
         """Wait for sensor data to be ready"""
         start = monotonic()
         while not self._data_ready_bit:
